Add search functionality to the Search Page
- Convert the text input field into a controlled component that
uses the searchText state
- Create an event handler, handleSearchText, that updates the
searchText state whenever there is a change in the search field.
- Supply a callback as the second argument to this.setState inside
handleSearchText. This callback will fetch books from the API and
update the books state.
This looks hacky, but the alternative options considered for
solving this problem looked even worse. The main complication occurs
when passing an up to date search text (this.state.searchText) to
BooksAPI.search(). The second callback in this.setState is only run
after state is updated, and that is why it works.
- componentDidUpdate was one of the options that was tried and it
also worked but it kept repeating the BooksAPI request no matter
what conditional expression I could think of as suggested in the
docs